token_checks.py
"""
Simple token safety checks placeholder.
"""

from web3 import Web3

def is_contract_verified(token_address: str) -> bool:
    # Placeholder: always return True
    # In real use, integrate blockchain explorer API to check verification status
    return True

def advanced_safety_checks(token_address: str) -> dict:
    """
    Perform a series of safety checks on the token contract.
    Returns a dict with check names mapping to bool results.
    """
    checks = {
        "Contract verified": is_contract_verified(token_address),
        # Add more checks as needed, e.g. 'Honeypot': not is_honeypot(token_address)
    }
    return checks
